1.8/gke: use k8s API instead of kubectl to check existence of test namespace
With dlespiau/kube-test-harness#38 the helper to get namespaces via the
k8s is available from kube-test-harness. Use it to check for the
existence of the ciliumNamespace test namespace and error out if it
already exists.

We could attempt to delete it, but this might lead to bad results if
pointed against the wrong cluster.

Signed-off-by: Tobias Klauser <tobias@isovalent.com>