1.8/gke: Correct number of pods in Cilium
The number of expected pods depends on the number of nodes in the
cluster. The Cilium and nodeinit daemonsets will both have one pod on
each node, and the Operator will have as many pods as replicas defined
in its deployment. For the Operator, our GKE manifest has one replica.

Signed-off-by: Chris Tarazi <chris@isovalent.com>